services/order_generator/requirements.txt
kafka-python==2.0.2
boto3==1.34.0
aws-msk-iam-sasl-signer-python==1.0.1
